Add script to generate benchmark datasets
This script takes in a set of gene sequences, removes duplicates,
clusters them based on Mash sketch distances, and for each large
enough cluster partitions the sequences into a set for use in the
graph and a set for benchmarking alignment.

By clustering the sequences first, we can control the estimated
minimum distance (and thus estimated alignment score) for each
data set.